# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.Translation:
    properties:
      destination:
        example: en
        type: string
      original:
        example: текст для перевода
        type: string
      source:
        example: auto
        type: string
      translation:
        example: text for translation
        type: string
    type: object
...
  entity.TranslationHistory:
    properties:
      history:
        items:
          $ref: '#/definitions/entity.Translation'
        type: array
    type: object
  entity.User:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      role:
        example: user
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      username:
        example: johndoe
        type: string
    type: object
  entity.UserAccount:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      onboarding_required:
        example: true
        type: boolean
      preferences:
        $ref: '#/definitions/entity.UserPreferences'
      profile:
        $ref: '#/definitions/entity.UserProfile'
      role:
        example: user
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      username:
        example: johndoe
        type: string
    type: object
  entity.UserActivity:
    properties:
      actor_email:
        example: admin@example.com
        type: string
      actor_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      client_ip:
        type: string
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      email:
        example: editor@example.com
        type: string
      error_code:
        example: invalid_role
        type: string
      event:
        example: role_change
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      new_role:
        example: editor
        type: string
      old_role:
        example: user
        type: string
      status:
        example: success
        type: string
      user_agent:
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  entity.UserPreferences:
    properties:
      arabic_level:
        example: basic
        type: string
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      daily_goal_minutes:
        example: 15
        type: integer
      fallback_langs:
        example:
        - id
        - en
        items:
          type: string
        type: array
      interests:
        example:
        - tafsir
        - hadith
        items:
          type: string
        type: array
      preferred_content_lang:
        example: id
        type: string
      preferred_ui_lang:
        example: id
        type: string
      quran_recitation_id:
        type: string
      quran_translation_source_id:
        type: string
      reader_mode:
        example: arabic_translation
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  entity.UserProfile:
    properties:
      country_code:
        example: ID
        type: string
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      display_name:
        example: John
        type: string
      onboarding_completed_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      onboarding_version:
        example: 1
        type: integer
      personalization_enabled:
        example: true
        type: boolean
      timezone:
        example: Asia/Jakarta
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  v1.AdminUserActivityList:
    properties:
      activity:
        items:
          $ref: '#/definitions/entity.UserActivity'
        type: array
      total:
        example: 42
        type: integer
    type: object
  v1.AdminUserList:
    properties:
      total:
        example: 42
        type: integer
      users:
        items:
          $ref: '#/definitions/entity.UserAccount'
        type: array
    type: object
...
  v1.Error:
    properties:
      code:
        example: invalid_request_body
        type: string
      details: {}
      error:
        example: message
        type: string
      message:
        example: message
        type: string
      request_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      retry_after:
        example: 60
        type: integer
    type: object
...
  v1.SetUserRole:
    properties:
      email:
        example: editor@example.com
        type: string
      role:
        example: editor
        type: string
    required:
    - email
    - role
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Surau classical book reader API
  title: Go Clean Template API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      description: Admin-only paginated user management list. Supports q, role, and
        email_verified filters.
      operationId: admin-list-users
      parameters:
      - description: Search username, email, or display name
        in: query
        name: q
        type: string
      - description: Role filter
        enum:
        - user
        - editor
        - admin
        in: query
        name: role
        type: string
      - description: Email verification status
        in: query
        name: email_verified
        type: boolean
      - default: 50
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.AdminUserList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: List admin users
      tags:
      - admin
  /admin/users/{id}:
    get:
      description: Admin-only user account detail by user ID.
      operationId: admin-get-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserAccount'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Get admin user detail
      tags:
      - admin
  /admin/users/{id}/activity:
    get:
      description: Admin-only role-change audit history for one user.
      operationId: admin-get-user-activity
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - default: 50
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.AdminUserActivityList'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Get admin user activity
      tags:
      - admin
  /admin/users/role:
    patch:
      consumes:
      - application/json
      description: Admin-only user role assignment for user, editor, or admin.
      operationId: admin-set-user-role
      parameters:
      - description: User role update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.SetUserRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Set user role
      tags:
      - admin
...
  /translation/history:
    get:
      description: Show all translation history for current user
      operationId: history
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.TranslationHistory'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Show history
      tags:
      - translation
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"